create component for time range dropdown
